STYLE: controller page divide
